requirements.txt
psycopg2-binary==2.9.9
yfinance==0.2.36
pandas
numpy
python-dotenv
lxml
jupyter

matplotlib